added logging for the validator

+++ frameworks/fmi2/src/main/java/org/intocps/maestro/framework/fmi2/Fmi2FmuValidator.java
@@ -6,15 +6,20 @@ import com.fujitsu.vdmj.VDMJ;
 import com.fujitsu.vdmj.VDMSL;
 import com.fujitsu.vdmj.lex.Dialect;
 import com.fujitsu.vdmj.runtime.Interpreter;
+import com.fujitsu.vdmj.values.BooleanValue;
+import com.fujitsu.vdmj.values.Value;
 import fmi2vdm.FMI2SaxHandler;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
 import org.intocps.fmi.IFmu;
+import org.intocps.maestro.ast.LexToken;
 import org.intocps.maestro.core.messages.IErrorReporter;
 import org.intocps.maestro.framework.fmi2.vdm.annotations.ast.ASTOnFailAnnotation;
 import org.intocps.maestro.framework.fmi2.vdm.annotations.in.INOnFailAnnotation;
 import org.intocps.orchestration.coe.FmuFactory;
 import org.intocps.orchestration.coe.modeldefinition.ModelDescription;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.xml.parsers.SAXParser;
 import javax.xml.parsers.SAXParserFactory;
@@ -26,20 +31,22 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class Fmi2FmuValidator implements IFmuValidator {
+    final static Logger logger = LoggerFactory.getLogger(Fmi2FmuValidator.class);
 
     @Override
     public boolean validate(String id, URI path, IErrorReporter reporter) {
         try {
+            logger.trace("Validating: {} at {}", id, path);
             SAXParserFactory factory = SAXParserFactory.newInstance();
             SAXParser saxParser = factory.newSAXParser();
             FMI2SaxHandler handler = new FMI2SaxHandler(path.getPath(), "var");
             IFmu fmu = FmuFactory.create(null, path);
-            //check schema
-            ModelDescription md = new ModelDescription(fmu.getModelDescription());
+            //check schema. The constructor checks the schema
+            new ModelDescription(fmu.getModelDescription());
 
+            logger.trace("Generate VDM specification for: {}", id);
             //convert to VDM
             saxParser.parse(fmu.getModelDescription(), handler);
-
 
             String vdm = handler.getFMIModelDescription().toVDM("\t");
 
@@ -65,6 +72,7 @@ public class Fmi2FmuValidator implements IFmuValidator {
                                     "ModelVariables_2.2" + ".7.vdmsl", "TypeDefinitions_2.2.3.vdmsl", "UnitDefinitions_2.2.2.vdmsl",
                                     "VariableNaming_2.2.9.vdmsl", "VendorAnnotations_2.2.6.vdmsl"};
 
+                    logger.trace("Copying static standard specification for id: {}", id);
                     List<File> specFiles = Arrays.stream(fmi2StaticModelFiles).map(specPath -> {
                         File tmp;
                         try {
@@ -87,27 +95,40 @@ public class Fmi2FmuValidator implements IFmuValidator {
 
                     specFiles.add(modelSped);
                     synchronized (INOnFailAnnotation.class) {
-                        controller.parse(specFiles);
-                        if (controller.typeCheck() == ExitStatus.EXIT_OK) {
+                        if (controller.parse(specFiles) == ExitStatus.EXIT_OK) {
+                            if (controller.typeCheck() == ExitStatus.EXIT_OK) {
 
-                            INOnFailAnnotation.failures.clear();
-                            Interpreter interpreter = controller.getInterpreter();
-                            interpreter.init();
-                            interpreter.execute("isValidFMIModelDescription(var)");
+                                INOnFailAnnotation.failures.clear();
+                                Interpreter interpreter = controller.getInterpreter();
+                                logger.trace("Initialize VDM interpreter and execute validation for id: {}", id);
+                                interpreter.init();
+                                Value result = interpreter.execute("isValidFMIModelDescription(var)").deref();
 
-                            boolean success = INOnFailAnnotation.failures.isEmpty();
-                            INOnFailAnnotation.failures.forEach(msg -> reporter.warning(0, msg + " URI: " + path, null));
+                                boolean success = false;
+                                if (result instanceof BooleanValue) {
+                                    success = result.boolValue(null);
+                                }
 
-                            //clean up
-                            INOnFailAnnotation.failures.clear();
+                                boolean hasAnnotations = !INOnFailAnnotation.failures.isEmpty();
+                                logger.trace("Specification for id '{}', compliant = {}, annotation = {}", id, success, hasAnnotations);
+                                INOnFailAnnotation.failures
+                                        .forEach(msg -> reporter.warning(0, msg, new LexToken(path.toString() + "/modelDescription" + ".xml", 0, 0)));
 
-                            for (File specFile : specFiles) {
-                                FileUtils.deleteQuietly(specFile);
+                                //clean up
+                                INOnFailAnnotation.failures.clear();
+
+                                for (File specFile : specFiles) {
+                                    FileUtils.deleteQuietly(specFile);
+                                }
+                                //FIXME we need to have better control over this
+                                return true;
+                            } else {
+                                logger.trace("Specification for id '{}' did not type check", id);
+                                reporter.warning(999, "Internal error could not check spec for: " + path, null);
+                                return false;
                             }
-                            //FIXME we need to have better control over this
-                            return true;
                         } else {
-
+                            logger.trace("Specification could not parse id '{}'", id);
                             reporter.warning(999, "Internal error could not check spec for: " + path, null);
                             return false;
                         }
